AppCodes/requirements.txt
flask
flask-sqlalchemy
flask_login
psycopg2-binary
email_validator
gunicorn
supabase
requests
google-generativeai
gunicorn
python-dotenv
